Initialise catalog db, add error handling in APIs

+++ src/catalog_server/views.py
@@ -1,52 +1,23 @@
+from flask import request, jsonify, Response
 from sqlalchemy import exc
 
-from flask import Flask, request, jsonify, Response
-from flask_sqlalchemy import SQLAlchemy
-from flask_marshmallow import Marshmallow
-import os
-
-app = Flask(__name__)
-basedir = os.path.abspath(os.path.dirname(__file__))
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///' + os.path.join(basedir, 'crud.sqlite')
-db = SQLAlchemy(app)
-ma = Marshmallow(app)
-
-
-class Book(db.Model):
-    id = db.Column(db.Integer, primary_key=True, autoincrement = True)
-    title = db.Column(db.VARCHAR)
-    topic = db.Column(db.VARCHAR)
-    count = db.Column(db.Integer)
-    __table_args__ = (
-        db.CheckConstraint('count >= 0'),
-        {})
-
-    def __init__(self, title, topic, count):
-        self.title = title
-        self.topic = topic
-        self.count = count
-
-
-class BookSchema(ma.Schema):
-    class Meta:
-        # Fields to expose
-        fields = ('id', 'title', 'topic', 'count')
-
+from src.catalog_server.models import Book, BookSchema, db, app
 
 book_schema = BookSchema()
 books_schema = BookSchema(many=True)
 
 
-# TODO error handling
+# TODO error handling: catch more specific exceptions?
 # Endpoint to create a new book item
 @app.route("/books", methods=["POST"])
 def add_book():
     # fixme Add validation here?
     new_book = Book(request.json['title'], request.json['topic'], request.json['count']['value'])
-
-    db.session.add(new_book)
-    db.session.commit()
-
+    try:
+        db.session.add(new_book)
+        db.session.commit()
+    except:
+        return Response(status=500, mimetype='application/json')
     return jsonify(book_schema.dump(new_book))
 
 
@@ -55,18 +26,20 @@ def add_book():
 def get_books():
     filtered_books = Book.query.filter_by(topic=request.args.get('topic'))
     result = books_schema.dump(filtered_books)
-    return jsonify(result)  # fixme remove .data
+    return jsonify(result)
 
 
 # Endpoint to get details of a book given the id
 @app.route("/books/<id>", methods=["GET"])
 def book_detail(id):
-    return book_schema.jsonify(Book.query.get(id))
+    book = Book.query.get(id)
+    if not book:
+        return Response(f"Book with id {id} not found.", status=404, mimetype='application/json')
+    return book_schema.jsonify(book)
 
 
 # Endpoint to update details of a book
 @app.route("/books/<id>", methods=["PATCH"])
-# fixme PATCH?
 def book_update(id):
     book = Book.query.get(id)
     book.topic = request.json.get('topic') or book.topic
@@ -83,18 +56,22 @@ def book_update(id):
     try:
         db.session.commit()
     except exc.IntegrityError:
-        return Response(f"0 books remaining. Cannot {request.json['count'].get('_operation')}.", status=400, mimetype='application/json')
+        return Response(f"0 books remaining. Cannot {request.json['count'].get('_operation')}.", status=400,
+                        mimetype='application/json')
     return book_schema.jsonify(book)
 
 
 # Endpoint to delete a book
 @app.route("/books/<id>", methods=["DELETE"])
 def book_delete(id):
-    # TODO error handling: delete non-existent book
     book = Book.query.get(id)
-    db.session.delete(book)
-    db.session.commit()
-
+    if not book:
+        return Response(f"Book with id {id} not found.", status=404, mimetype='application/json')
+    try:
+        db.session.delete(book)
+        db.session.commit()
+    except:
+        return Response(status=500, mimetype='application/json')
     return book_schema.jsonify(book)
